requirements.txt
requests>=2.31.0
eth-account>=0.11.0
eth-abi>=5.0.0
python-dotenv>=1.0.0
